Only use Debian pool if base_ntp_server is not defined
If base_ntp_server is defined, aka we have set a custom ntp server, we
do not want to use the Debian pool.

Ansible Molecule verifier checks the expected files.
The `pool` line needs to be removed, as we now expect the Debian
pool line to be missing if base_ntp_server is defined.

Closes: #20