Debian: add support for trixie-updates repository
We usually hardcode the release name, to provide according support
information in the related notes. Because we didn't explicitly check for
trixie so far (which will be released as new Debian stable release as of
2025-08-09), the trixie-updates repository was not yet available (even
if base_enable_stable_updates is enabled).

Since there's no official support information available for trixie yet.
Use the generic ${ansible_distribution_release}-updates, to avoid ending
up without an *-updates entry. Once the support information for trixie
will be available, we can add an explicit check for "trixie".